Added size check to Wrapper of Client and server and moved ReuseableBuffer to its own file
